dilferd fix?

+++ index.js
@@ -13,11 +13,7 @@ const client = new Client({
     GatewayIntentBits.GuildMessageReactions,
     GatewayIntentBits.MessageContent,
   ],
-  partials: [
-    Partials.Message,
-    Partials.Channel,
-    Partials.Reaction,
-  ],
+  partials: [Partials.Message, Partials.Channel, Partials.Reaction],
 });
 
 const getUserRequestFrequency = (() => {
@@ -41,7 +37,7 @@ const isThrottled = (userId, messageTime) => {
     return false;
   }
 
-  const lockoutTime = 2 ** (userRequestFrequency.messageCount) * 1_000;
+  const lockoutTime = 2 ** userRequestFrequency.messageCount * 1_000;
   const timeSinceLastMessage = messageTime - userRequestFrequency.lastMessageTime;
 
   if (timeSinceLastMessage < lockoutTime) {
@@ -63,19 +59,13 @@ client.once(Events.ClientReady, () => {
 
 client.on(Events.MessageCreate, async (message) => {
   if (message.author.bot) return;
-
-  if (!message.content.match(/br[aue]h/ig)) return;
-
+  if (!message.content.match(/br[aue]h/gi)) return;
   if (isThrottled(message.author.id, message.createdTimestamp)) return;
-
-  const bruhEmoji = message.guild?.emojis.cache.find((emoji) => emoji.name === 'bruh');
-  const response = await message.channel.send(bruhEmoji ? `bruh ${bruhEmoji}` : 'bruh');
-  if (!bruhEmoji) return;
   try {
-    await Promise.all([
-      message.react(bruhEmoji),
-      response.react(bruhEmoji),
-    ]);
+    const response = await message.channel.send('bruh');
+    const bruhEmoji = message.guild?.emojis.cache.find((emoji) => emoji.name === 'bruh');
+    if (!bruhEmoji) return;
+    await Promise.all([message.react(bruhEmoji), response.react(bruhEmoji)]);
   } catch (err) {
     console.error(err);
   }
@@ -95,7 +85,9 @@ client.on(Events.MessageCreate, async (message) => {
   if (message.author.bot) return;
   if (message.author.id !== DILFERD_ID) return;
 
+  if (!message.content.match(/^\s*[n]/i)) return;
   if (!isThrottled(message.author.id, message.createdTimestamp)) return;
+
   try {
     await message.delete();
   } catch (err) {
